+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Checksum calculation method changed to GitHub API SHA256 from MD5
+
 ## [1.22.1] - 2025-06-05
 ### Added
 - Developed release notes for experimental GPU metadata repo